trans_path_web/requirements.txt
fastapi
uvicorn
asyncpg
haversine
python-dotenv